+++ pages/src/localdev.ts
@@ -50,12 +50,14 @@ import './esl-media-demo/test-media-source';
 
 import {ESLDemoSidebar} from './navigation/navigation';
 import {ESLDemoMarquee} from './landing/landing';
+import {ESLDemoSearchBox} from './navigation/header/header-search';
 
 ESLVSizeCSSProxy.observe();
 
 // Register Demo components
 ESLDemoSidebar.register();
 ESLDemoMarquee.register();
+ESLDemoSearchBox.register();
 
 // Register ESL Components