+++ src/modules/esl-panel-group/core/esl-panel-group.ts
@@ -1,7 +1,8 @@
 import {ExportNs} from '../../esl-utils/environment/export-ns';
-import {attr, ESLBaseElement} from '../../esl-base-element/core';
+import {attr, jsonAttr, ESLBaseElement} from '../../esl-base-element/core';
 import {afterNextRender} from '../../esl-utils/async/raf';
 import {bind} from '../../esl-utils/decorators/bind';
+import {format} from '../../esl-utils/misc/format';
 import {memoize} from '../../esl-utils/decorators/memoize';
 import {CSSClassUtils} from '../../esl-utils/dom/class';
 import {ESLMediaRuleList} from '../../esl-media-query/core';
@@ -24,6 +25,8 @@ export class ESLPanelGroup extends ESLBaseElement {
 
   /** Rendering mode of the component (takes values from the list of supported modes; 'accordion' by default) */
   @attr({defaultValue: 'accordion'}) public mode: string;
+  /** Rendering mode class pattern. Uses {@link format} syntax for `mode` placeholder */
+  @attr({defaultValue: 'esl-{mode}-view'}) public modeCls: string;
   /** Element {@link TraversingQuery} selector to add class that identifies the rendering mode (ESLPanelGroup itself by default) */
   @attr({defaultValue: ''}) public modeClsTarget: string;
   /** Class(es) to be added during animation ('animate' by default) */
@@ -32,13 +35,14 @@ export class ESLPanelGroup extends ESLBaseElement {
   @attr({defaultValue: 'auto'}) public fallbackDuration: number | 'auto';
   /** List of comma-separated "modes" to disable collapse/expand animation (for both Group and Panel animations) */
   @attr() public noCollapse: string;
-
   /**
    * Define accordion behavior
    * `single` allows only one Panel to be open.
    * `multiple` allows any number of open Panels.
    * */
   @attr({defaultValue: 'single'}) public accordionGroup: string;
+  /** Action params to pass into panels when executing reset action (happens when mode is changed) */
+  @jsonAttr({defaultValue: {noCollapse: true}}) public transformParams: PanelActionParams;
 
 
   /** Height of previous active panel */
@@ -46,26 +50,11 @@ export class ESLPanelGroup extends ESLBaseElement {
   /** Fallback setTimeout timer */
   protected _fallbackTimer: number = 0;
 
-  static get observedAttributes() {
+  static get observedAttributes(): string[] {
     return ['mode', 'accordion-group'];
   }
 
-  /** ESLMediaRuleList instance of the mode mapping */
-  @memoize()
-  public get modeRules() {
-    return ESLMediaRuleList.parse(this.mode);
-  }
-
-  /** @returns current mode */
-  public get currentMode(): string {
-    return this.modeRules.activeValue || '';
-  }
-
-  protected get actionParams(): PanelActionParams {
-    return {initiator: 'group', activator: this};
-  }
-
-  protected connectedCallback() {
+  protected connectedCallback(): void {
     super.connectedCallback();
     this.bindEvents();
 
@@ -73,14 +62,14 @@ export class ESLPanelGroup extends ESLBaseElement {
     this.updateMode();
   }
 
-  protected disconnectedCallback() {
+  protected disconnectedCallback(): void {
     super.disconnectedCallback();
     this.modeRules.removeListener(this._onModeChange);
 
     this.unbindEvents();
   }
 
-  protected attributeChangedCallback(attrName: string, oldVal: string, newVal: string) {
+  protected attributeChangedCallback(attrName: string, oldVal: string, newVal: string): void {
     if (!this.connected || oldVal === newVal) return;
     if (attrName === 'mode') {
       this.modeRules.removeListener(this._onModeChange);
@@ -97,7 +86,7 @@ export class ESLPanelGroup extends ESLBaseElement {
     }
   }
 
-  protected bindEvents() {
+  protected bindEvents(): void {
     this.addEventListener('esl:before:show', this._onBeforeShow);
     this.addEventListener('esl:show', this._onShow);
     this.addEventListener('esl:before:hide', this._onBeforeHide);
@@ -105,7 +94,7 @@ export class ESLPanelGroup extends ESLBaseElement {
     this.addEventListener('transitionend', this._onTransitionEnd);
   }
 
-  protected unbindEvents() {
+  protected unbindEvents(): void {
     this.removeEventListener('esl:before:show', this._onBeforeShow);
     this.removeEventListener('esl:show', this._onShow);
     this.removeEventListener('esl:before:hide', this._onBeforeHide);
@@ -113,40 +102,72 @@ export class ESLPanelGroup extends ESLBaseElement {
     this.removeEventListener('transitionend', this._onTransitionEnd);
   }
 
-  /** Update element state according to current mode */
-  protected updateMode() {
-    this.setAttribute('view', this.currentMode);
+  /** Updates element state according to current mode */
+  protected updateMode(): void {
+    const prevMode = this.getAttribute('current-mode');
+    const currentMode = this.currentMode;
+    this.setAttribute('current-mode', currentMode);
+    // TODO: @deprecated will be removed with the 4th esl version
+    this.setAttribute('view', currentMode);
+
+    this.updateModeCls();
+    this.reset();
+
+    if (prevMode !== currentMode) {
+      this.$$fire('change:mode', {detail: {prevMode, currentMode}});
+    }
+  }
+
+  /** Updates mode class marker */
+  protected updateModeCls(): void {
+    const {modeCls, currentMode} = this;
+    if (!modeCls) return;
     const $target = TraversingQuery.first(this.modeClsTarget, this);
     if (!$target) return;
     ESLPanelGroup.supportedModes.forEach((mode) => {
-      $target.classList.toggle(`esl-${mode}-view`, this.currentMode === mode);
+      const className = format(modeCls, {mode});
+      $target.classList.toggle(className, currentMode === mode);
     });
-
-    this.reset();
   }
 
-  /** @returns Panels that are processed by the current panel group */
+  /** @returns ESLMediaRuleList instance of the mode mapping */
+  @memoize()
+  public get modeRules(): ESLMediaRuleList<string> {
+    return ESLMediaRuleList.parse(this.mode);
+  }
+
+  /** @returns current mode */
+  public get currentMode(): string {
+    return this.modeRules.activeValue || '';
+  }
+
+  /** @returns panels that are processed by the current panel group */
   public get $panels(): ESLPanel[] {
     const els = Array.from(this.querySelectorAll(ESLPanel.is));
     return els.filter((el) => this.includesPanel(el)) as ESLPanel[];
   }
 
-  /** @returns Panels that are active */
-  public get $activePanels() {
+  /** @returns panels that are active */
+  public get $activePanels(): ESLPanel[] {
     return this.$panels.filter((el: ESLPanel) => el.open);
   }
 
-  /** @returns Whether the collapse/expand animation should be handheld by the group */
-  public get shouldCollapse() {
+  /** @returns whether the collapse/expand animation should be handheld by the group */
+  public get shouldCollapse(): boolean {
     const noCollapseModes = this.noCollapse.split(',').map((mode) => mode.trim());
     return !noCollapseModes.includes('all') && !noCollapseModes.includes(this.currentMode);
   }
 
-  /** @returns Action params config that is used by controlled {@link ESLPanel}s */
+  /** @returns action params config that's used (inherited) by controlled {@link ESLPanel}s */
   public get panelConfig(): PanelActionParams {
     return {
       noCollapse: !this.shouldCollapse || (this.currentMode === 'tabs')
     };
+  }
+
+  /** @returns merged panel action params for show/hide requests from the group */
+  protected mergeActionParams(...params: PanelActionParams[]): PanelActionParams {
+    return Object.assign({initiator: 'group', activator: this}, ...params);
   }
 
   /** Condition-guard to check if the passed target is a Panel that should be controlled by the Group */
@@ -155,35 +176,35 @@ export class ESLPanelGroup extends ESLBaseElement {
     return target.$group === this;
   }
 
-  /** Show all panels besides excluded ones */
-  public showAll(excluded: ESLPanel[] = []) {
-    this.$panels.forEach((el) => !excluded.includes(el) && el.show(this.actionParams));
+  /** Shows all panels besides excluded ones */
+  public showAll(excluded: ESLPanel[] = [], params: PanelActionParams = {}): void {
+    this.$panels.forEach((el) => !excluded.includes(el) && el.show(this.mergeActionParams(params)));
   }
-  /** Hide all active panels besides excluded ones */
-  public hideAll(excluded: ESLPanel[] = []) {
-    this.$activePanels.forEach((el) => !excluded.includes(el) && el.hide(this.actionParams));
+  /** Hides all active panels besides excluded ones */
+  public hideAll(excluded: ESLPanel[] = [], params: PanelActionParams = {}): void {
+    this.$activePanels.forEach((el) => !excluded.includes(el) && el.hide(this.mergeActionParams(params)));
   }
-  /** Toggle all panels by predicate */
-  public toggleAllBy(shouldOpen: (panel: ESLPanel) => boolean) {
-    this.$panels.forEach((panel) => panel.toggle(shouldOpen(panel), this.actionParams));
+  /** Toggles all panels by predicate */
+  public toggleAllBy(shouldOpen: (panel: ESLPanel) => boolean, params: PanelActionParams = {}): void {
+    this.$panels.forEach((panel) => panel.toggle(shouldOpen(panel), this.mergeActionParams(params)));
   }
 
-  /** Reset to default state applicable to the current mode */
-  public reset() {
+  /** Resets to default state applicable to the current mode */
+  public reset(): void {
     ESLPanel.registered.then(() => {
-      if (this.currentMode === 'open') this.toggleAllBy(() => true);
+      if (this.currentMode === 'open') this.toggleAllBy(() => true, this.transformParams);
       if (this.currentMode === 'tabs' || (this.currentMode === 'accordion' && this.accordionGroup === 'single')) {
         const $activePanel = this.$panels.find((panel) => panel.initiallyOpened);
-        this.toggleAllBy((panel) => panel === $activePanel);
+        this.toggleAllBy((panel) => panel === $activePanel, this.transformParams);
       }
       if (this.currentMode === 'accordion' && this.accordionGroup === 'multiple') {
-        this.toggleAllBy((panel) => panel.initiallyOpened);
+        this.toggleAllBy((panel) => panel.initiallyOpened, this.transformParams);
       }
     });
   }
 
-  /** Animate the height of the component */
-  protected onAnimate(from: number, to: number) {
+  /** Animates the height of the component */
+  protected onAnimate(from: number, to: number): void {
     const hasCurrent = this.style.height && this.style.height !== 'auto';
     if (hasCurrent) {
       this.style.height = `${to}px`;
@@ -200,18 +221,18 @@ export class ESLPanelGroup extends ESLBaseElement {
   }
 
   /** Pre-processing animation action */
-  protected beforeAnimate() {
+  protected beforeAnimate(): void {
     CSSClassUtils.add(this, this.animationClass);
   }
 
   /** Post-processing animation action */
-  protected afterAnimate() {
+  protected afterAnimate(): void {
     this.style.removeProperty('height');
     CSSClassUtils.remove(this, this.animationClass);
   }
 
-  /** Init a fallback timer to call post-animate action */
-  protected fallbackAnimate() {
+  /** Inits a fallback timer to call post-animate action */
+  protected fallbackAnimate(): void {
     const time = +this.fallbackDuration;
     if (isNaN(time) || time < 0) return;
     if (this._fallbackTimer) clearTimeout(this._fallbackTimer);
@@ -220,7 +241,7 @@ export class ESLPanelGroup extends ESLBaseElement {
 
   /** Process {@link ESLPanel} pre-show event */
   @bind
-  protected _onBeforeShow(e: CustomEvent) {
+  protected _onBeforeShow(e: CustomEvent): void {
     const panel = e.target;
     if (!this.includesPanel(panel)) return;
     if (this.currentMode === 'accordion' && this.accordionGroup === 'multiple') return;
@@ -229,7 +250,7 @@ export class ESLPanelGroup extends ESLBaseElement {
 
   /** Process {@link ESLPanel} show event */
   @bind
-  protected _onShow(e: CustomEvent) {
+  protected _onShow(e: CustomEvent): void {
     const panel = e.target;
     if (!this.includesPanel(panel)) return;
     if (this.currentMode !== 'tabs') return;
@@ -244,7 +265,7 @@ export class ESLPanelGroup extends ESLBaseElement {
 
   /** Process {@link ESLPanel} pre-hide event */
   @bind
-  protected _onBeforeHide(e: CustomEvent) {
+  protected _onBeforeHide(e: CustomEvent): void {
     // TODO: refactor
     if (this.currentMode === 'open') {
       e.preventDefault();
@@ -255,9 +276,9 @@ export class ESLPanelGroup extends ESLBaseElement {
     this._previousHeight = this.offsetHeight;
   }
 
-  /** Catching CSS transition end event to start post-animate processing */
+  /** Catches CSS transition end event to start post-animate processing */
   @bind
-  protected _onTransitionEnd(e?: TransitionEvent) {
+  protected _onTransitionEnd(e?: TransitionEvent): void {
     if (!e || e.propertyName === 'height') {
       this.afterAnimate();
     }
@@ -265,7 +286,7 @@ export class ESLPanelGroup extends ESLBaseElement {
 
   /** Handles mode change */
   @bind
-  protected _onModeChange() {
+  protected _onModeChange(): void {
     this.updateMode();
   }
 }