+++ src/polyfills/es5-target-shim.ts
@@ -1,6 +1,6 @@
 // Shim for modern browsers with ES6 class syntax support
 // Shim based on https://github.com/webcomponents/polyfills/blob/master/packages/custom-elements/ts_src/custom-elements.ts
-export function shimES5ElementConstructor(BuiltInHTMLElement: typeof HTMLElement) {
+export function shimES5ElementConstructor(BuiltInHTMLElement: typeof HTMLElement): void {
   if (
     // No Reflect, no classes, no need for shim because native custom elements require ES2015 classes or Reflect.
     window.Reflect === undefined || window.customElements === undefined ||